requirements.txt
functions-framework==3.*
google-auth-httplib2
oauth2client
google-api-python-client>=2.105.0
google-auth>=2.27.0
google-cloud-storage>=2.14.0